Enable custom state adapters
